Remove array_cast option from tco_migration_flags list
Summary: I don't believe the `array_cast` option in the `tco_migration_flags` list is being used to control anything in Hack anymore. This diff simply removes it from the list of options. I am leaving the empty list in case people want to use it in the future. I'll put up a diff if people want to commandeer to remove the `tco_migration_flags` list entirely.

Reviewed By: jamesjwu

Differential Revision: D33673571

fbshipit-source-id: 4c1600024b5cc25caf96ecf931ab7d30bd47c3a9