Output timing info as well
Summary:
This makes it easier to do performance measurements.

Note that we won't capture the time taken to dump the actual trace output - I
wanted to still dump the metadata at the top of the file.

I also changed the filename scheme to include the request start time so
multiple requests to the same endpoint will now produce multiple files.

Reviewed By: dkgi

Differential Revision: D33779681

fbshipit-source-id: 93e5e8ba0c2cf1dbea70db9fa6f899a3f4b7b8d2