Replace use of facts parser with decl parser in OCaml FFI
Summary:
Now that the new facts has been [rolled out to 100%](https://www.internalfb.com/diff/D33688656), it might be a good time to start preparing for replacement of the old facts parser.

To that end, this diff
- Renames `extract_facts_as_json_ffi0` to `extract_facts_as_json_ffi`. This is the FFI accessed by the [OCaml code](https://www.internalfb.com/code/fbsource/[0bc043f095f3ecc418bc16c5c71ba43de9c2d938]/fbcode/hphp/hack/src/facts/facts_parser.ml?lines=38)
- Within that function, replaces `facts_parser::from_text` with `direct_decl_parser::parse_decls_without_reference_text` + `facts::Facts::facts_of_decls`
- Removes 2 unused FFIs, `extract_facts_ffi0` and `facts_to_json`

In later diffs we'll refactor out further dependencies on the facts parser and eventually delete it.

Reviewed By: shayne-fletcher

Differential Revision: D33721383

fbshipit-source-id: bbf7845f00590dad24fd4aa7f9fec90fe01d2d15