Commit exclusion list for pessimisation tests and enable test
Summary:
This uses the `--disabled-extension` flag of verify.py to declare that `.pess_disabled` turns off a test for the `typecheck:validate_pessimisation` target.

- `.pess_exp` are tests that we want to differ under pessimisation
- `.pess_disabled` are tests we want to follow up on / fix

This diff re-enables the validate_pessimisation test target

Follow up diff will update verify.py so if a test is diasabled, run it anyway and compare it to the .exp, and report that it can be re-enabled if the exp matches.

Differential Revision: D33752279

fbshipit-source-id: 43fde2cc634deba981c322ad907ccfac5847050b