removing stream_errors flag: remove branching on flag + dead code
Summary:
Error streaming has now been rolled out everywhere, this diff stack cleans up the legacy  code. This first diff remove the branching on the flag plus some of the dead code.

The SUBSCRIBE_DIAGNOSTIC command is also not used with error streaming so we delete it and all associated code.

Reviewed By: ljw1004

Differential Revision: D33635223

fbshipit-source-id: bcec5181e1a76bc289e231ac30447e54226ef303