Simplify intersections with Tany in subtyping
Summary:
When we see
```
_ & t1 <: t2
```
simplify it to
```
_ <: t2
```
This is to prevent issues with disjunctions where we end up doing `t1 <: t2`.

A common source of such subtype assertions is `is C<_>` tests on variables of type `Tany`.

The use of `Tany` is of course unsound and unpredictable; sound dynamic will provide a long term solution. But in the meantime, we want to remove sources of brittleness.

Reviewed By: vsiles

Differential Revision: D33851431

fbshipit-source-id: c1a1e9b3153f9946e8ec588a65fa76c18ee3b563