Avoid unnecessary disjunction in subtyping
Summary:
We were generating unnecessary disjunctions during subtyping for the special case of
```
t1 <: ?t2
```
where `t1` is a `newtype`, generic parameter, or expression-dependent type, with upper bound.

In the case that it's syntactically obvious that `null` isn't in `t1` (by looking at the bound), we can reduce this to `t1 <: t2` instead of the more general `t1 <: t2 || b <: ?t2` where `b` is the bound.

Differential Revision: D33841664

fbshipit-source-id: f6e8c4c78e4732b3a683e258bd6898e70d94c519