Don't require default construction for types with BlobDecoder
Summary:
To decode a type using BlobDecoder, you need to first default
construct the type, then pass it into the decoder. This can be
problematic if the type isn't default constructable. To support this
use case, add a new function to BlobDecoder called make() (and
makeWhole()). Instead of deserializing into an existing value, these
deserialize and return a new value. If the type has a static method
called makeForSerde(), this is used to construct the value. Otherwise
the type is default constructed, as before.

While I'm making modifications, add support to BlobEncoder/BlobDecoder
to support std::map since I'll need that.

Reviewed By: oulgen

Differential Revision: D33872445

fbshipit-source-id: a6de3ad6887f66e05c4ee25b7278590c0ced98e8