Add macro for unit cache affecting flags
Summary:
Instead of having the list in getUnitCacheBreakingOptions(), provide a
macro to enumerate the list. I want to serialize these options and I
can use such a macro to automatically generate the code without having
to keep it in sync.

Since this is only being used by unit-cache (again), move it back
there.

Reviewed By: oulgen

Differential Revision: D33812062

fbshipit-source-id: 41144561347afe3963208605786a615d614be87b