PosId<P> -> Positioned<S, P>
Summary:
PosId maps to the ocaml (pos, symbol) concept. It currently hardcodes `Symbol`,
and we'd like [the option] to introduce a family of typesafe new-type wrappers
for Symbol, like Classname, Methodname, etc. So it seems useful to surface
the type of symbol, and decouple position types from symbol types.

To pave the way, make PosId generic on the symbol, and rename to Positioned.

Reviewed By: shayne-fletcher

Differential Revision: D33901687

fbshipit-source-id: e6ea2c4b0bbc69d250d2c3c0bc59990ff0b0b4d9