Remove dune build of hh_single_compile executable
Summary:
We are migrating to `hh_single_compile.cpp` as this dune build is just calling rust code and our replacement does everything `hh_single_compile` does and more.
It's also more difficult to support c++ dependencies on when introducing more cxx ffis.

[RFC] Do developers use this and how can we make it easier to use `hh_single_compile.cpp` instead?

There do not appear to be any more uses of the dune build of `hh_single_compile`.

Reviewed By: vsiles

Differential Revision: D33855892

fbshipit-source-id: efa3bf12794a5c0bb8fe18ce98ceceb65bb8c1b0