Share attribute definition between HHVM and HackC (#8964)
Summary:
Pull Request resolved: https://github.com/facebook/hhvm/pull/8964

A bit of a precursor to the new HhasProgram parser.

Currently in the bytecode_printer, we set `HhasClassFlags::IS_SEALED` by doing a string comparison on `"__Sealed"` and then check this flag and emit `"sealed"` as a part of the attribute list. HHVM then parses this string list and does a string match on `"sealed"`. We really don't need this many representations of the `Sealed` attribute.

If we only use one source of truth for attributes:
- we don't need to do all these extra string comparisons
- when we add a new feature, we only need to add a new attribute in one place

NOTE:This is very lightweight. Its just linking a single header file with no dependencies.

RFC: This makes the *raw* hhas less verbose. The HHVM team uses `-vEval.DumpHhas=1` to see hhas and this hhas remains unchanged. This diff does affect the *raw* output hhas. If the Hack team uses the raw hhas instead of -`vEval.DumpHhas=1`  to debug, I can add a "to_string()" in the bytecode_printer as well.

Next diff will be adding some asserts on these attributes. Want to put this one up for feedback on ^.

Reviewed By: shayne-fletcher

Differential Revision: D33715471

fbshipit-source-id: 7b49205f33a85379cd555939e73ca2e0f5860b7e