refine linter_switch_check w.r.t. labels
Summary:
Label `===` only care about the label name, type is not always available so, it
is not enforced.
For example, consider
```
enum class E : I {
  C A = ...
}
enum class F : J {
  D A = ...
}
```
Even if they don't have the same type, `E#A === F#A` because they really just are the string `"A"` to the runtime.
This has the side effect to trigger this linter during switch, even in genuine scenarios:
```
function f(Label<E, I> $label) {
  switch ($label) {
   case E#A: ...
  }
}
```
In this case, it will comment that `E#A: Label<E, C>` doesn't have the exact same type as `$label: Label<E, I>` (because `Label` is invariant in its second type parameter), but really, this is fine.

 This diff makes sure the linter doesn't fire in such cases.

Reviewed By: andrewjkennedy

Differential Revision: D33707711

fbshipit-source-id: 30f45ae90970309684b11858582898702d1aa16b