Clean up variadics part 1: add variadic flag to fun_type
Summary:
Our current representation of variadic functions is complex and error-prone. Eventually I'd like to remove `fun_arity` by including the variadic parameter in the ordinary parameter list, and simply setting a bit in function type flags to indicate that the last parameter is a variadic.

First step is to introduce the flag. It should be the case that this flag is set if and only if the arity is `Fvariadic _`.

Differential Revision: D33914773

fbshipit-source-id: ae8f22ce11d8ebdb75f583af715005436a6963f4