Replace jit::type with DataType in DecRefProfile
Summary:
The change to add jit::type to DecRefProfile increased its size by 32 bytes (DecRefProfile went from 16 bytes to 48 bytes).

However, the profiled type is only used when it is a DataType. So replace jit::type with DataType enum (int8_t), to drop the size of DecRefProfile back to 20 bytes.

Reviewed By: mofarrell

Differential Revision: D33837532

fbshipit-source-id: ad7433479001d2c2967e5dd77cfc9af89007b628