Clean up and simplify redundant checks in assembler
Summary:
Couple cleanups I found while implementing a hhas-parser for the HhasProgram struct HackC gives us instead of the bytecode string:

Assembler errors in as.cpp should be purely for when we fail to read in the bytecode string given to us from HackC. Checks on the types of the things we read in were already done in the parser in HackC and are redundant. We should make them verifier errors for now.

Checks on which attributes are allowed should go with the rest of these checks in the HackC parser.

Reviewed By: ricklavoie

Differential Revision: D32710982

fbshipit-source-id: f86f1082069e655b53237fdcc21c204ccaf544c6