LazyUnitContentsLoader shouldn't copy string if provided
Summary:
LazyUnitContentsLoader can either receive a path (and attempt lazy
loading), or be given a string up front. Right now if given a string,
it copies the string into its own internal buffer. Soon I'll be using
the "given string" case more and don't wish to copy the string
everytime. Refactor LazyUnitContentsLoader to not copy the
string. Instead if just stores a StringPiece to the given string (it's
the responsibility of the caller to keep the string alive).

Reviewed By: oulgen

Differential Revision: D33929526

fbshipit-source-id: ee6886e1e31be10fbc34cc5b88d353491ed77e41