Split RepoOptions into RepoOptionsFlags and RepoOptions
Summary:
The parsing code path only cares about a subset of the information in
RepoOptions. Basically just the runtime options and their associated
hash. Split RepoOptions into RepoOptionsFlags, which just stores this
information, and RepoOptions, which contains a RepoOptionsFlags and
the rest.

Update the relevant code paths to use RepoOptionsFlags instead. The
advantage of all this is that when we distribute the parsing steps,
there's less information we need to transmit.

Reviewed By: oulgen

Differential Revision: D33929386

fbshipit-source-id: 1ebe768c723ec40ec5f675dfb32732d2a99523a6