Match `popen` with `pclose`
Summary:
`popen` is a Linux system call that opens a new process and returns a file descriptor. Its manual page specifies that "the return value from `popen()` is a normal standard I/O stream in all respects save that it must be closed with `pclose()` rather than `fclose(3)`." More recent versions of Clang, such as Clang 12, warn when a file descriptor returned by `popen` is closed with `fclose`, which led me to this error.

See also:
* `popen(3)`: https://man7.org/linux/man-pages/man3/popen.3.html
* `fclose(3)`: https://man7.org/linux/man-pages/man3/fclose.3.html

Reviewed By: ricklavoie

Differential Revision: D33844086

fbshipit-source-id: cd09b613f01c9f97b89e9fa13cd47c8c38131ccb