Increase buffer size for platform010 `-Wformat-truncation=`
Summary:
When built with Clang 12, the compiler emits a warning `-Wformat-truncation=` for the `neo_hdf` target:
```
hphp/neo/neo_hdf.c: In function '_hdf_read_string':
hphp/neo/neo_hdf.c:1383:48: warning: 'snprintf' output may be truncated before the last format character [-Wformat-truncation=]
 1383 |             snprintf(fullpath, PATH_MAX, "%s/%s", pwd, name);
      |                                                ^
hphp/neo/neo_hdf.c:1383:13: note: 'snprintf' output 2 or more bytes (assuming 4097) into a destination of size 4096
 1383 |             snprintf(fullpath, PATH_MAX, "%s/%s", pwd, name);
      |             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
```

The issue appears to be that `pwd`'s size may be 4096 characters, whereas the write size is also 4096, despite also including a `/` followed by another format argument for `name`.

My naive suggestion, not being a maintainer of this code and not understanding it well, would be to increase the size of the buffer so as to allow room for the entirety of `pwd` to be written.

Reviewed By: ricklavoie

Differential Revision: D33828913

fbshipit-source-id: 5a3e2e51b3ba07e46f178c0b5aaadb8095fd697e