Add runtime option for inline configs
Summary:
Add a `TaintConfigurationJson` runtime option to make it easier to specify configs when running on servicelab. This is a temporary hack until we set up an fbpkg for these configs that can be loaded by servicelab.

For the same reason, also create the taint output directory if not already present.

Reviewed By: dkgi

Differential Revision: D33899580

fbshipit-source-id: f80040b9ae3478b49948e6a04829bd910aeb0597