Avoid allocations in iopPreamble
Summary:
For larger opcode names (>16 bytes) we need to make a heap allocation
as SSO can no longer kick in. We end up allocating and freeing a string each
time we hit this, only to not do anything with it if logging is turned off.

I saw this with `iopFCallClsMethodD` which is 18 chars but should show up with
other stuff too

Reviewed By: oulgen

Differential Revision: D33853469

fbshipit-source-id: a02c342d857b97a0bac12882fd6c13dfea261460