Add HH\Readonly\Shapes class
Summary:
Since we aren't planning on supporting readonly polymorphism (due to the increase in complexity to the language of adding such a feature), we'll likely want some interim Readonly\ namespace versions of common functions in Shapes. These don't pollute the main systemlib namespace, and mostly just call directly into their original Shapes class counterparts, which to me feel like the least invasive way to add this without introducing too much technical debt/cognitive load.

We can incrementally remove these if we do decide to add polymorphism for readonly, but right now the tradeoffs of doing so would certainly be too much complexity without TIC. So we should expect these functions to be around for a while. Feedback and concerns appreciated!

Reviewed By: oulgen

Differential Revision: D33730938

fbshipit-source-id: 3948731bb9055de7378ed35e058596e5de307f34