Fix chgrp error handling
Summary:
Our error handling for chgrp isn't correct if the group does
not exist. We assume that getgrnam_r will return 0 if the group
doesn't exist. However the documentation says it can fail in a variety
of ways. If the buffer passed in is too small, it will return ERANGE
and you need to increase the buffer. Furthermore, if the group doesn't
exist, it may return 0, or it is allowed to return ENOENT or
ESRCH. Handle these cases.

Reviewed By: mofarrell

Differential Revision: D33984915

fbshipit-source-id: d1efe6fcea6bbf2807e40fa61bdf185a163e4699