Fix binding of ImplicitContext::activeCtx inside hphp_process_init
Summary:
Inside hphp_process_init, we check RO::EvalEnableImplicitContext, and
if true, bind ImplicitContext::activeCtx into RDS. The problem with
this is that we haven't loaded any RepoGlobalData out of the repo file
yet. So if we're in RepoAuthoritative mode, and the RepoGlobalData
specifies EvalEnableImplicitContext (and the command-line doesn't), we
won't bind the handle, and probably crash later when we try to use it.

The clearest way this manifests is by trying to run any of the
implicit context unit tests with --repo-separate. It crashes almost
immediately because we don't bind the handle, because the
RepoGlobalData isn't loaded yet (and that's the *only* way to set
runtime options with stand-alone HHBBC).

The easiest fix would to just move the bind to after the
g_vmProcessInit call, which loads RepoGlobalData. However, we must
bind the handle before calling jit::mcgen::processInit (which creates
the unique stubs). We can't move the call to g_vmProcessInit any
earlier because thatmust be run after jit::mcgen::processInit. So we
have a circular dependency. The easiest way to fix the dependency is
just bind the handle unconditionally. If the option isn't actually
set, we'll just waste 8 bytes in RDS, which isn't worth worrying
about.

Reviewed By: oulgen

Differential Revision: D33997900

fbshipit-source-id: c63796e62cb7ef57499656212899d3eed3ba2b4e