Stop retaining collection types with singletons
Summary:
First introduced in D32271255 (https://github.com/facebook/hhvm/commit/d49cafb849768c6d439883886a7d71fb5f5edee2) in order for the direct decl parser to produce identical facts as the facts parser. With this behavior, a collection type with a singleton (e.g. `vec[1]`) gets parsed but that with multiple items (e.g. `vec[1, 2]`) are ignored.

It seems removal of this special treatment for singleton collections would not affect autoloading. This diff removes it to simplify the decl parser.

Differential Revision: D33674404

fbshipit-source-id: c8fd3e42e38f340239604e3a60dde1ff4133067f