Rename PathAndHash to PathAndOptionalHash
Summary:
`PathAndHash` doesn't always have a hash. The `m_hash` field is optional.

There's another `PathAndHash` type defined in Facts-land, inside of `AutoloadDB`. It's one of the return types for `AutoloadDB`'s methods:

https://www.internalfb.com/code/fbsource/[deb5fe2f6d9138ad8fd7e3e57890be07c0ab88af]/fbcode/hphp/runtime/ext/facts/autoload-db.h?lines=272

Within `AutoloadDB`, it's not obvious which `PathAndHash` structure we're referring to. This diff renames `HPHP::Facts::PathAndHash` to `HPHP::Facts::PathAndOptionalHash` to make that more clear.

Reviewed By: ricklavoie

Differential Revision: D33959187

fbshipit-source-id: 139a0b6156a6a74a7f83297fd08e549daf679934