fix clippy::ptr_arg lint
Summary:
The fact that aast::Block is a type alias for Vec<Stmt> sometimes leads
to valid lint, when a function takes `&Block` as an argument, because
an argument of type `&Vec<T>` is overly restrictive and should be a slice `&[T]`.

I went on a quest to formalize the the definition of Block : `struct Block(Vec<Stmt>)`,
and started fixing up the affected code. Some of it was good, but it ended up
being too much churn & boilerplate so I abandoned the idea. I'm glad I tried
because I learned my way around oxidized (somewhat).

Instead this diff directly fixes the lint.

Differential Revision: D34030007

fbshipit-source-id: a1ddb55b8c6eb66f38790ace0d37cfef5319e068