remove redundant lower bounds in disjunctive typing constraints
Summary:
When we have an intersection on the left of subtyping, a disjunctive constraint is created. Since the solver doesn't do back-tracking, it choses the first disjunct that doesn't immediately fail. However, that might not be the best choice. This implements a simplification phase to recognise a special case where one disjunct is strictly better than another, and remove the worse one. Concretely, if we have `t1 <: #1` and `t2 <: #1` and `t1 <: t2`, then we always want to choose the `t1` constraint, so we remove the `t2` one.

This situation is somewhat rare, since when we build the intersection, we already check for subtyping. However, it can occur when an operation, e.g., array indexing, reveals a subtyping relationship. For example, if `$s : shape('f' => string)` and we have `$s as KeyedContainer<_,_>`, then `$s : shape('f' => string) & KeyedContainer<arraykey, mixed>`. If we then do `$s['f']` we can get a lower bound of `mixed` and one of `string`, and we never want to used the `mixed` one.

Reviewed By: andrewjkennedy

Differential Revision: D34006343

fbshipit-source-id: 8531646a24fb202c24e67abb77c889119f15fe5f