Allow readonly return without return type hint
Summary:
Allow our parser to parse the readonly keyword after a colon without a return type specifier.

This is kind of annoying because our parsing of type specifiers is based around the presence of the colon. But introducing an _opt version of parse_type_specifier is a reasonable way of handling it.

Reviewed By: KendallHopkins

Differential Revision: D33994503

fbshipit-source-id: 9102ed90d15b04147877ead8fb2639cbcfc251c4