Refactor AutoloadDB return types
Summary: `AutoloadDB`'s methods return a bizarre mix of `std::tuple`, `std::pair`, and structs. I'm refactoring this so we define and return structs everywhere within `AutoloadDB`. I'm also refactoring callsites to avoid destructuring assignment when we return structs, so that callsite correctness doesn't depend on the order in which fields are defined.

Reviewed By: ricklavoie

Differential Revision: D33959936

fbshipit-source-id: 3d8d77453cb59c541413f32cdd51440896f37f03