AutoloadDB::RowIter -> AutoloadDB::MultiResult
Summary:
I'm refactoring native Facts to work with a networked DB, instead of being tied to SQLite. `RowIter` is a lazy generator which wraps a query object and extracts values from the DB on demand. It exposes a `SQLiteQuery` to the caller, which blocks us from creating an implementation of `AutoloadDB` that doesn't use `SQLiteQuery`.

This diff refactors `AutoloadDB::RowIter` to `AutoloadDB::MultiResult`, which uses a closure to hide its implementation details from the caller.

Reviewed By: ricklavoie

Differential Revision: D33960913

fbshipit-source-id: f7768d482c6afc8054492b9bada7a066c5142ad1