Rename AutoloadDB::analyze to AutoloadDB::runPostBuildOptimizations
Summary:
SQLite has an [`ANALYZE` command](https://www.sqlite.org/lang_analyze.html) which takes a while to run but massively improves some of the query plans in the native Facts DB. If we've just built a Facts DB from scratch, we run it once after building.

I'm abstracting `AutoloadDB` so I can back it with a networked DB like MySQL. I'm not sure if there even is any sort of `ANALYZE` equivalent in MySQL, but I'm renaming the function so it's a bit more generic, and letting the docblock explain the SQLite-ness of the function. I'm also moving the logging into `AutoloadDB`'s implementation, where it will become the SQLite-specific implementation.

Also, this function was in a weird place in the file so I'm moving the method down a bit.

Reviewed By: bdchik

Differential Revision: D33961035

fbshipit-source-id: 9d03b9c78e5223ccdae1619f00e4551a243a4456