Remove forking from HPHPc
Summary:
HPHPc, by default, forks before executing. The parent just
waits for the child to do all the work. This is an old old holdover
form when HPHPc was a transpiler and we needed to fork before
executing g++ (to save memory for some reason). This is all
unnecessary, so drop the forking.

There was an option (--nofork) to disable forking, which is now
unnecessary, so remove it and the few uses of it.

Reviewed By: oulgen

Differential Revision: D34054005

fbshipit-source-id: c3f3e6ed10ee05c19cd55fb3ca7a6069470f7862