Remove reliance on error ordering
Summary:
The function `Errors.try_with_error f g` first performs the computation `f` then inspects the errors generated as a side effect.

If there are no errors, the computation `g` is performed, its result returned and its side-effecting errors retained.

If there _were_ errors resulting from `f`, it selects the error that was added last and retains irrespective of what happens in `g`

The retention of an error from a failing `f` cannot in general be discard but in the position this function is used here, we are guaranteed that if `f` fails, we generate an error in `g`.

In the next diff, we move to (mostly) pure errors which will peturb error ordering which will cause a change in test results simply because of that change in ordering.

By using `Errors.try_` here we eliminate the reliance on error ordering meaning the isolating the effects of subsequent changes.

Reviewed By: vsiles

Differential Revision: D34005542

fbshipit-source-id: b19602dc6800301c5113bbc8447c8d92cb406bc0