Improve subtyping wrt constraint types
Summary: The subtyping cases for newtype and dependent on the left weren't enabled when the type on the right was a constraint type. This is needlessly inflexible, since those cases just move to using the upper bound on the left. This fixes that with an aim to using constraint types for array index operations.

Reviewed By: andrewjkennedy

Differential Revision: D34075604

fbshipit-source-id: 4e51f514b6cf49dd003ce7c307e9643e29cca7e4