Fix race condition in async MySQL extension
Summary:
abandon() is an optimization to short-circuit an async event if we can't get it started on a processing thread. This method can only be called from the web request thread that initiated that event, because it manipulates request-local data (including potentially destroying the abandoned event's Hack WaitHandle object).

This async MySQL processing thread code calls abandon() after the event has started. Instead, it should deliver the failure to the web request thread by marking the event finished and throwing an exception on unserialize. I've already checked that we throw when query_op->ok() is false, so, assuming that query_op is not ok() here, we can just mark the event finished.

Reviewed By: jano

Differential Revision: D34058230

fbshipit-source-id: 674172329ae05ef2ac3b512891239f10e1463a37