Replace DBData with AutoloadDB::Handle in SymbolMap
Summary:
`DBData` is a key describing the location and read-only state of a SQLite DB. We use it as a hashmap key, and as a means of connecting to a SQLite DB in the filesystem. But it contains SQLite implementation details, so we need to abstract it out to a common interface if we're going to make `SymbolMap` compatible with both SQLite and a networked DB.

In this case, the common interface I've chosen is a factory function object which returns a DB connection.

Reviewed By: ricklavoie

Differential Revision: D33991294

fbshipit-source-id: e028185dcddf157a43478b697d4d468a1259e8d2