Prepare for CtxConstants
Summary:
The `HhasProgram/HackCUnit` struct that is passed from HackC to HHVM currently passes `CtxConstants` with a coeffects field `Pair(Slice<Ctx>, Slice<Str>)`.

`bytecode_printer/print.rs` converts the `Slice<Ctx>` to strings and concatenates with the `String<Str>` to generate the hhas.

In HHVM, when the `hackc-translator` sees the `Slice<Ctx>`, it needs a way to convert to strings as well.

There are a couple options:
0. Just make HackC emit a `Slice<Str>` for both of them. This was simplest and the other options were not convincingly better. (This is what is implemented in this diff)

1. Make `pub enum Ctx` and `to_string()` method part of a cxx ffi that is shared. This looked like it would disorganize most of the rust coeffects module.

2. Write another `Ctx -> String` method in HHVM. Not great for maintability.

3. Share some `StaticCoeffects/RuntimeCoeffects` type from HHVM to HackC and use that instead of `Ctx` (similar to Attr). This looked more complicated than sharing `Attr` and I didn't want to attempt this before better understanding that code.

Reviewed By: oulgen

Differential Revision: D34080853

fbshipit-source-id: eacbb188958013789e980d7cc0ccea337b6d75f7