Replace redundant `NAST_check_error`
Summary:
We currently raise a `NAST_check_error`, `Magic` (3011), from `typing_object_get` when we encounter a `__construct` method access. This causes problems when moving to pure errors since we would have to deal with apply `Reasons_callback.t` to NAST check errors in subtyping.

To avoid changing behaviour, it's clear that we need to raise _some_ error when `__construct` is encountered to ensure our failure condition is maintained when doing `Errors.try_` or equivalent (i.e. at least one error is raised) but we are able to change the _specific_ error we raise in this case. This is beause we will _already_ raise this error in `illegal_name_check`.

This diff removes the NAST check error and instead raises a new typing error `ConstructNotInstanceMethod` for this case. NB The original 3011 is still raised so this new error will appear in addition to this. There is some redundancy in these errors but there isn't an obvious way of avoiding it.

Differential Revision: D34043777

fbshipit-source-id: a8a105b157cdf027a7ff91b1ae8a27f5f335c110