Bound the number of profiled array elements
Summary: We'd like to keep profile data updates O(1) wherever possible. In particular, we can restrict value-level profiling (that we use to target the "shallow-release" operation) to a bounded number of elements. In most cases, an array with more than 16 elements is some kind of monotyped array that we can treat homogeneously. We can also bail early in this loop as soon as we find any refcounted element.

Differential Revision: D34092451

fbshipit-source-id: 226ed0afc67003273b69df708c32bd6bd16f6873