Make TypingDeclProvider a trait
Summary: Abstract over the implementation of TypingDeclProvider, allowing for completely unrelated implementations of class-decl-folding. This trait serves the same purpose as the Typing_classes_heap API in hh_server: for folded decl, it abstracts over the fact that member types are stored separately from folded class decls. It also abstracts over the concrete type of a folded class decl. Like Typing_classes_heap, it does *not* abstract over the type of a folded class element (`ClassElt`).

Reviewed By: shayne-fletcher

Differential Revision: D34087646

fbshipit-source-id: b403d50bd2287c8df7b85faca3ea0904e92f7c98