Separate decl-folding logic from FoldedDeclProvider
Summary:
Extract decl-folding logic into a new type called DeclFolder, making it clear that most of this logic does not interact with a ShallowDeclProvider or the FoldedDeclProvider's cache of folded decls.

This also makes it possible to reuse the decl-folding logic in other implementations of FoldedDeclProvider. For HHVM, we might have an implementation of FoldedDeclProvider whose methods all call into C++ procedures which do work similar to what LazyFoldedDeclProvider does after this diff: collect the folded decls of the parents and the shallow decl of the class, then call (back into Rust via FFI) into DeclFolder::decl_class to produce the folded decl, then store it in cache and return it.

Reviewed By: shayne-fletcher

Differential Revision: D34092806

fbshipit-source-id: 52cfeaf169c81ce3962ff975372b1090d502c83b