I think we don't need commit_batch
Summary:
These are the ONLY PLACES in the entire codebase that use "commit/revert on the stack of local changes" in SharedMem.ml. I find that a hard-to-understand mechanism, and I believe this use of the mechanism isn't needed, so I think we can get rid of it...

I believe the significance of these lines is that we've decided not to use the "batch" mechanism at all.

Let's spell it out. Prior to this diff, if "stop_at_errors" was true, then if there were parse-errors in any IDE files then we'd refrain from committing those errorfull versions into Ast_provider or Fixme_provider. Moreover, if there were parse-errors in any files disk or IDE, then we'd refrain from committing those changes into File_provider.

The intent of this was that even if there were parse errors, well, the various heaps would still have "known good previous versions" in them. Kasper spelled out a motivation in D3854040 (https://github.com/facebook/hhvm/commit/4d6bb20991d334033961d764b8edff976f87fa9d): "
We don't want mid-typing parse error to trigger invalidation of large swaths of dependent declarations. To achieve this, for IDE files we don't update parse trees if they had errors (so there will be no change to declarations)."

I assume he specifically meant that if we kept the last-known-good Ast around, then even while there's a parse error in a decl, we'd end up decl-diffing against the last-known-good Ast and therefore wouldn't find any decl-diffs and therefore wouldn't compute any big fanout.

(Is that really true? I wonder what happens to fanout calculation now when you're in the IDE and you typed a character which invalidates a declaration? Does it do decl-diffing and find that the declaration has changed? Does it calculate a big fanout? What stops it from checking that big fanout? -- jakebailey says that with shallow fanout we no longer do phase 2. So what happens to fanout in this case must be controlled by the logic in ServerTypeCheck.LazyCheck.)

Anyway, after this diff, all that goes out the window. Let's spell out what happens now:

(1) Line 290 pushes the stack for Files, Asts, Fixmes:
```
  if stop_at_errors then (
    File_provider.local_changes_push_sharedmem_stack ();
    Ast_provider.local_changes_push_sharedmem_stack ();
    Fixme_provider.local_changes_push_sharedmem_stack ()
  );
```

(2) Line 340 commits those stacks for all changes to ide and disk files. (Ide and disk files are just a split/partition of "to_check".) The only thing we don't commit is `File_provider.commit_batch disk_files.`. NOTE: the only changes we ever expect to see in File, Ast and Fixme are those that came from ide/disk files!
```
  if stop_at_errors then (
    File_provider.local_changes_commit_batch ide_files;
    Ast_provider.local_changes_commit_batch ide_files;
    Fixme_provider.local_changes_commit_batch ide_files;
    Ast_provider.local_changes_commit_batch disk_files;
    Fixme_provider.local_changes_commit_batch disk_files;
```

(3) Line 347 removes the stack:
```
... if stop_at_errors then ( ...
	    File_provider.local_changes_pop_sharedmem_stack ();
    Ast_provider.local_changes_pop_sharedmem_stack ();
    Fixme_provider.local_changes_pop_sharedmem_stack ();
```

If we're creating a stack, and then committing every single change in our stack, and then abandoning the stack -- then there was no point even having a stack in the first place.

NOTE... The one wildcard here is about File_provider. We're deliberately not keeping any disk_files that we rechecked in the File_provider. I assume this is either to save space, or because of an oversight when it was first introduced. But in any case, it should all be immaterial, since the File_heap is going away.

Differential Revision: D33543406

fbshipit-source-id: fc554e0ffeda49c2fc9750c890a0d48981cf1503