Refactor and optimize how we track special functions
Summary:
This diff started with the intention to track all functions we've seen
(to enable dumping coverage at the end of the run) but I realized we can also
make some more optimizations while we're at it.

The comments in the code should explain *how* the new data structure works. I'd
appreciate feedback inline if that can be clarified.

As for the *why*: We need to store some info to see if a function was hit.
Now that we need to use extra memory here anyway to store at least a bit per
function, we can use that to our advantage and skip all the eviction and most
of the hashmap lookup overhead.

This also is extensible for the future when we may need to store whether a
function has a special model (e.g. for builtins) - we could continue to store
that cheaply.

The special function tracking was moved back to being a global thing (instead
of per request) as that allows for an improved cache hit rate. The usages in
`interpreter.cpp` still call into this via the `State` object though -
`Configuration::get()` returns a `shared_ptr` and atomic refcounting was
incredibly hot in local testing. We can use the `State` object to keep the
pointer alive and avoid atomics each time we need to look up a function this way.

In practice I'm seeing another ~10-15% win in local testing in terms of
wall clock request performance via this diff.

I chose to put the code here and not on the `Func` object itself for easier
separation of concerns - the rest of the HHVM code stays cleaner.

Reviewed By: dkgi

Differential Revision: D34054447

fbshipit-source-id: 2d4608e1b9ef10c0690ad249a9769a92f3de4cbf