add module interfaces
Summary:
Refactoring of the glean hack indexer.

1. Define an interface `.mli` file for each module, except for `symbol_builder_types` which contains only type declarations.
2. Replace some "open statements" with qualified names. Change naming convention accordingly, for instance ` add_container_decl_fact` is renamed `container_decl` and with the module prefix becomes `Add_fact.container_decl`.

See here for motivation https://fb.workplace.com/groups/403065016408786/permalink/1631947443520531/

This is a first step of refactoring. The next step will be to regroup a (abstract) type and the functions which operate on it  in the same module, rather than put all the types open in `symbol_builder_types` and having various modules that use these types.

Reviewed By: nhawkes

Differential Revision: D34093589

fbshipit-source-id: 89e00778275ab9ff3dff634c5372d5ee29bf87d7