Move character location header to runtime/base
Summary:
The next diff gets rid of everything in the hphp/parser directory -
location.h is the only file that needs to be preserved from that project for
some error message use cases. A better place for it is in hphp/runtime/base. (Not deleting hphp/parser/location.h in this diff because it's used in scanner.h)

Reviewed By: oulgen

Differential Revision: D33932875

fbshipit-source-id: 06c11ecc66854e00d05077eb0ea7037299be49dc